+++ CHANGES.rst
@@ -62,6 +62,10 @@ Bug Fixes
 
 - ``astropy.nddata``
 
+- ``astropy.sphinx``
+
+  - Added slightly more useful debug info for AstropyAutosummary. [#2024]
+
 - ``astropy.stats``
 
 - ``astropy.table``